+++ apps/frontend/src/utils/constants.ts
@@ -60,7 +60,7 @@ export const servicesConfig = {
 export const EXPORT_RECORD_LIMIT = 500;
 export const DEFAULT_TIMEOUT_SECONDS = 5000;
 export const LIQUIDATION_RESERVE_AMOUNT = 20;
-export const GAS_LIMIT_BLOCK = 6_800_000;
+export const MAX_GAS_LIMIT = 6_800_000;
 export const GAS_LIMIT_OPEN_TROVE = 1_000_000;
 export const GAS_LIMIT_ADJUST_TROVE = 650_000;
 export const GAS_LIMIT_STABILITY_POOL = 500_000;
